Revert "Plan 9 from Bell Labs Fourth Edition 2007-04-14"
This reverts commit 223a035810d484657ee1a815f68faa8211009d6d.

+++ sys/src/cmd/cwfs/9p1.c
@@ -37,7 +37,7 @@ mkchallenge(Authinfo *aip)
 {
 	int i;
 
-	srand((uintptr)aip + time(nil));
+	srand((ulong)aip + time(nil));
 	for(i = 0; i < CHALLEN; i++)
 		aip->chal[i] = nrand(256);
